my-ai-platform/scripts/requirements.txt
# Scripts 전용 의존성 (로컬 실행용)
# 데이터 임포트, 동기화, 테스트 스크립트용

# Core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# LLM
langchain-ollama>=0.2.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.13.0

# Graph Database
neo4j>=5.16.0

# Vector Database
pymilvus>=2.4.0

# Embeddings (Milvus 동기화용)
sentence-transformers>=3.3.0
torch>=2.5.0
transformers>=4.47.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Utilities
python-dotenv>=1.0.1
